requirements.txt
langchain
langchain-google-genai
python-dotenv
google-generativeai
attrs>=21.2.0